requirements.txt
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
jupyter>=1.0.0
ipykernel>=6.0.0
python-dotenv>=1.0.0
